+++ toss-spa/src/components/hireSection.ts
